Bench: 1310686

+++ src/Cuckoo.h
@@ -14,12 +14,12 @@ namespace cuckoo
 inline std::array<uint64_t, 8192> table;
 inline std::array<Move, 8192> move_table;
 
-inline uint64_t H1(uint64_t h)
+inline uint16_t H1(uint64_t h)
 {
     return (h >> 32) & 0x1fff;
 };
 
-inline uint64_t H2(uint64_t h)
+inline uint16_t H2(uint64_t h)
 {
     return (h >> 48) & 0x1fff;
 };
